Split some classes
